`analysis` -> `ragged` + `kinematics` (#291)
* analysis -> kinematics, ragged

* Bump minor version

* Remove mistakenly commited change

* top docstrings edits

---------
