+++ clouddrift/__init__.py
@@ -4,8 +4,9 @@ __version__ = version
 
 from clouddrift.raggedarray import RaggedArray
 import clouddrift.adapters
-import clouddrift.analysis
 import clouddrift.datasets
-import clouddrift.sphere
+import clouddrift.kinematics
+import clouddrift.ragged
 import clouddrift.signal
+import clouddrift.sphere
 import clouddrift.wavelet
