+++ clouddrift/ragged.py
@@ -1,16 +1,14 @@
 """
-This module provides common Lagrangian analysis and transformation
-functions.
+Transformational and inquiry functions for ragged arrays.
 """
 
 import numpy as np
-from typing import Optional, Tuple, Union, Iterable
+from typing import Tuple, Union, Iterable
 import xarray as xr
 import pandas as pd
 from concurrent import futures
 from datetime import timedelta
 import warnings
-from clouddrift.sphere import distance, bearing, position_from_distance_and_bearing
 
 
 def apply_ragged(
@@ -111,7 +109,7 @@ def apply_ragged(
             raise ValueError("The sum of rowsize must equal the length of arr.")
 
     # split the array(s) into trajectories
-    arrays = [unpack_ragged(np.array(arr), rowsize, rows, axis) for arr in arrays]
+    arrays = [unpack(np.array(arr), rowsize, rows, axis) for arr in arrays]
     iter = [[arrays[i][j] for i in range(len(arrays))] for j in range(len(arrays[0]))]
 
     # parallel execution
@@ -317,48 +315,6 @@ def prune(
     return ragged, rowsize
 
 
-def regular_to_ragged(
-    array: np.ndarray, fill_value: float = np.nan
-) -> tuple[np.ndarray, np.ndarray]:
-    """Convert a two-dimensional array to a ragged array. Fill values in the input array are
-    excluded from the output ragged array.
-
-    Parameters
-    ----------
-    array : np.ndarray
-        A two-dimensional array.
-    fill_value : float, optional
-        Fill value used to determine the bounds of contiguous segments.
-
-    Returns
-    -------
-    tuple[np.ndarray, np.ndarray]
-        A tuple of the ragged array and the size of each row.
-
-    Examples
-    --------
-    By default, NaN values found in the input regular array are excluded from
-    the output ragged array:
-
-    >>> regular_to_ragged(np.array([[1, 2], [3, np.nan], [4, 5]]))
-    (array([1., 2., 3., 4., 5.]), array([2, 1, 2]))
-
-    Alternatively, a different fill value can be specified:
-
-    >>> regular_to_ragged(np.array([[1, 2], [3, -999], [4, 5]]), fill_value=-999)
-    (array([1., 2., 3., 4., 5.]), array([2, 1, 2]))
-
-    See Also
-    --------
-    :func:`ragged_to_regular`
-    """
-    if np.isnan(fill_value):
-        valid = ~np.isnan(array)
-    else:
-        valid = array != fill_value
-    return array[valid], np.sum(valid, axis=1)
-
-
 def ragged_to_regular(
     ragged: Union[np.ndarray, pd.Series, xr.DataArray],
     rowsize: Union[list, np.ndarray, pd.Series, xr.DataArray],
@@ -409,10 +365,79 @@ def ragged_to_regular(
     :func:`regular_to_ragged`
     """
     res = fill_value * np.ones((len(rowsize), int(max(rowsize))), dtype=ragged.dtype)
-    unpacked = unpack_ragged(ragged, rowsize)
+    unpacked = unpack(ragged, rowsize)
     for n in range(len(rowsize)):
         res[n, : int(rowsize[n])] = unpacked[n]
     return res
+
+
+def regular_to_ragged(
+    array: np.ndarray, fill_value: float = np.nan
+) -> tuple[np.ndarray, np.ndarray]:
+    """Convert a two-dimensional array to a ragged array. Fill values in the input array are
+    excluded from the output ragged array.
+
+    Parameters
+    ----------
+    array : np.ndarray
+        A two-dimensional array.
+    fill_value : float, optional
+        Fill value used to determine the bounds of contiguous segments.
+
+    Returns
+    -------
+    tuple[np.ndarray, np.ndarray]
+        A tuple of the ragged array and the size of each row.
+
+    Examples
+    --------
+    By default, NaN values found in the input regular array are excluded from
+    the output ragged array:
+
+    >>> regular_to_ragged(np.array([[1, 2], [3, np.nan], [4, 5]]))
+    (array([1., 2., 3., 4., 5.]), array([2, 1, 2]))
+
+    Alternatively, a different fill value can be specified:
+
+    >>> regular_to_ragged(np.array([[1, 2], [3, -999], [4, 5]]), fill_value=-999)
+    (array([1., 2., 3., 4., 5.]), array([2, 1, 2]))
+
+    See Also
+    --------
+    :func:`ragged_to_regular`
+    """
+    if np.isnan(fill_value):
+        valid = ~np.isnan(array)
+    else:
+        valid = array != fill_value
+    return array[valid], np.sum(valid, axis=1)
+
+
+def rowsize_to_index(rowsize: Union[list, np.ndarray, xr.DataArray]) -> np.ndarray:
+    """Convert a list of row sizes to a list of indices.
+
+    This function is typically used to obtain the indices of data rows organized
+    in a ragged array.
+
+    Parameters
+    ----------
+    rowsize : list or np.ndarray or xr.DataArray
+        A list of row sizes.
+
+    Returns
+    -------
+    np.ndarray
+        A list of indices.
+
+    Examples
+    --------
+
+    To obtain the indices within a ragged array of three consecutive rows of sizes 100, 202, and 53:
+
+    >>> rowsize_to_index([100, 202, 53])
+    array([0, 100, 302, 355])
+    """
+    return np.cumsum(np.insert(np.array(rowsize), 0, 0))
 
 
 def segment(
@@ -513,516 +538,6 @@ def segment(
         return np.concatenate(segment_sizes)
 
 
-def rowsize_to_index(rowsize: Union[list, np.ndarray, xr.DataArray]) -> np.ndarray:
-    """Convert a list of row sizes to a list of indices.
-
-    This function is typically used to obtain the indices of data rows organized
-    in a ragged array.
-
-    Parameters
-    ----------
-    rowsize : list or np.ndarray or xr.DataArray
-        A list of row sizes.
-
-    Returns
-    -------
-    np.ndarray
-        A list of indices.
-
-    Examples
-    --------
-
-    To obtain the indices within a ragged array of three consecutive rows of sizes 100, 202, and 53:
-
-    >>> rowsize_to_index([100, 202, 53])
-    array([0, 100, 302, 355])
-    """
-    return np.cumsum(np.insert(np.array(rowsize), 0, 0))
-
-
-def position_from_velocity(
-    u: np.ndarray,
-    v: np.ndarray,
-    time: np.ndarray,
-    x_origin: float,
-    y_origin: float,
-    coord_system: Optional[str] = "spherical",
-    integration_scheme: Optional[str] = "forward",
-    time_axis: Optional[int] = -1,
-) -> Tuple[np.ndarray, np.ndarray]:
-    """Compute positions from arrays of velocities and time and a pair of origin
-    coordinates.
-
-    The units of the result are degrees if ``coord_system == "spherical"`` (default).
-    If ``coord_system == "cartesian"``, the units of the result are equal to the
-    units of the input velocities multiplied by the units of the input time.
-    For example, if the input velocities are in meters per second and the input
-    time is in seconds, the units of the result will be meters.
-
-    Integration scheme can take one of three values:
-
-        1. "forward" (default): integration from x[i] to x[i+1] is performed
-            using the velocity at x[i].
-        2. "backward": integration from x[i] to x[i+1] is performed using the
-            velocity at x[i+1].
-        3. "centered": integration from x[i] to x[i+1] is performed using the
-            arithmetic average of the velocities at x[i] and x[i+1]. Note that
-            this method introduces some error due to the averaging.
-
-    u, v, and time can be multi-dimensional arrays. If the time axis, along
-    which the finite differencing is performed, is not the last one (i.e.
-    x.shape[-1]), use the ``time_axis`` optional argument to specify along which
-    axis should the differencing be done. ``x``, ``y``, and ``time`` must have
-    the same shape.
-
-    This function will not do any special handling of longitude ranges. If the
-    integrated trajectory crosses the antimeridian (dateline) in either direction, the
-    longitude values will not be adjusted to stay in any specific range such
-    as [-180, 180] or [0, 360]. If you need your longitudes to be in a specific
-    range, recast the resulting longitude from this function using the function
-    :func:`clouddrift.sphere.recast_lon`.
-
-    Parameters
-    ----------
-    u : np.ndarray
-        An array of eastward velocities.
-    v : np.ndarray
-        An array of northward velocities.
-    time : np.ndarray
-        An array of time values.
-    x_origin : float
-        Origin x-coordinate or origin longitude.
-    y_origin : float
-        Origin y-coordinate or origin latitude.
-    coord_system : str, optional
-        The coordinate system of the input. Can be "spherical" or "cartesian".
-        Default is "spherical".
-    integration_scheme : str, optional
-        The difference scheme to use for computing the position. Can be
-        "forward" or "backward". Default is "forward".
-    time_axis : int, optional
-        The axis of the time array. Default is -1, which corresponds to the
-        last axis.
-
-    Returns
-    -------
-    x : np.ndarray
-        An array of zonal displacements or longitudes.
-    y : np.ndarray
-        An array of meridional displacements or latitudes.
-
-    Examples
-    --------
-
-    Simple integration on a plane, using the forward scheme by default:
-
-    >>> import numpy as np
-    >>> from clouddrift.analysis import position_from_velocity
-    >>> u = np.array([1., 2., 3., 4.])
-    >>> v = np.array([1., 1., 1., 1.])
-    >>> time = np.array([0., 1., 2., 3.])
-    >>> x, y = position_from_velocity(u, v, time, 0, 0, coord_system="cartesian")
-    >>> x
-    array([0., 1., 3., 6.])
-    >>> y
-    array([0., 1., 2., 3.])
-
-    As above, but using centered scheme:
-
-    >>> x, y = position_from_velocity(u, v, time, 0, 0, coord_system="cartesian", integration_scheme="centered")
-    >>> x
-    array([0., 1.5, 4., 7.5])
-    >>> y
-    array([0., 1., 2., 3.])
-
-    Simple integration on a sphere (default):
-
-    >>> u = np.array([1., 2., 3., 4.])
-    >>> v = np.array([1., 1., 1., 1.])
-    >>> time = np.array([0., 1., 2., 3.]) * 1e5
-    >>> x, y = position_from_velocity(u, v, time, 0, 0)
-    >>> x
-    array([0.        , 0.89839411, 2.69584476, 5.39367518])
-    >>> y
-    array([0.        , 0.89828369, 1.79601515, 2.69201609])
-
-    Integrating across the antimeridian (dateline) by default does not
-    recast the resulting longitude:
-
-    >>> u = np.array([1., 1.])
-    >>> v = np.array([0., 0.])
-    >>> time = np.array([0, 1e5])
-    >>> x, y = position_from_velocity(u, v, time, 179.5, 0)
-    >>> x
-    array([179.5      , 180.3983205])
-    >>> y
-    array([0., 0.])
-
-    Use the ``clouddrift.sphere.recast_lon`` function to recast the longitudes
-    to the desired range:
-
-    >>> from clouddrift.sphere import recast_lon
-    >>> recast_lon(x, -180)
-    array([ 179.5      , -179.6016795])
-
-    Raises
-    ------
-    ValueError
-        If u and v do not have the same shape.
-        If the time axis is outside of the valid range ([-1, N-1]).
-        If lengths of x, y, and time along time_axis are not equal.
-        If the input coordinate system is not "spherical" or "cartesian".
-        If the input integration scheme is not "forward", "backward", or "centered"
-
-    See Also
-    --------
-    :func:`velocity_from_position`
-    """
-    # Velocity arrays must have the same shape.
-    # Although the exception would be raised further down in the function,
-    # we do the check here for a clearer error message.
-    if not u.shape == u.shape:
-        raise ValueError("u and v must have the same shape.")
-
-    # time_axis must be in valid range
-    if time_axis < -1 or time_axis > len(u.shape) - 1:
-        raise ValueError(
-            f"time_axis ({time_axis}) is outside of the valid range ([-1,"
-            f" {len(x.shape) - 1}])."
-        )
-
-    # Input arrays must have the same length along the time axis.
-    if not u.shape[time_axis] == v.shape[time_axis] == time.shape[time_axis]:
-        raise ValueError(
-            f"u, v, and time must have the same length along the time axis "
-            f"({time_axis})."
-        )
-
-    # Swap axes so that we can differentiate along the last axis.
-    # This is a syntax convenience rather than memory access optimization:
-    # np.swapaxes returns a view of the array, not a copy, if the input is a
-    # NumPy array. Otherwise, it returns a copy. For readability, introduce new
-    # variable names so that we can more easily differentiate between the
-    # original arrays and those with swapped axes.
-    u_ = np.swapaxes(u, time_axis, -1)
-    v_ = np.swapaxes(v, time_axis, -1)
-    time_ = np.swapaxes(time, time_axis, -1)
-
-    x = np.zeros(u_.shape, dtype=u.dtype)
-    y = np.zeros(v_.shape, dtype=v.dtype)
-
-    dt = np.diff(time_)
-
-    if integration_scheme.lower() == "forward":
-        x[..., 1:] = np.cumsum(u_[..., :-1] * dt, axis=-1)
-        y[..., 1:] = np.cumsum(v_[..., :-1] * dt, axis=-1)
-    elif integration_scheme.lower() == "backward":
-        x[..., 1:] = np.cumsum(u_[1:] * dt, axis=-1)
-        y[..., 1:] = np.cumsum(v_[1:] * dt, axis=-1)
-    elif integration_scheme.lower() == "centered":
-        x[..., 1:] = np.cumsum(0.5 * (u_[..., :-1] + u_[..., 1:]) * dt, axis=-1)
-        y[..., 1:] = np.cumsum(0.5 * (v_[..., :-1] + v_[..., 1:]) * dt, axis=-1)
-    else:
-        raise ValueError(
-            'integration_scheme must be "forward", "backward", or "centered".'
-        )
-
-    if coord_system.lower() == "cartesian":
-        x += x_origin
-        y += y_origin
-    elif coord_system.lower() == "spherical":
-        dx = np.diff(x)
-        dy = np.diff(y)
-        distances = np.sqrt(dx**2 + dy**2)
-        bearings = np.arctan2(dy, dx)
-        x[..., 0], y[..., 0] = x_origin, y_origin
-        for n in range(distances.shape[-1]):
-            y[..., n + 1], x[..., n + 1] = position_from_distance_and_bearing(
-                y[..., n], x[..., n], distances[..., n], bearings[..., n]
-            )
-    else:
-        raise ValueError('coord_system must be "spherical" or "cartesian".')
-
-    return np.swapaxes(x, time_axis, -1), np.swapaxes(y, time_axis, -1)
-
-
-def velocity_from_position(
-    x: np.ndarray,
-    y: np.ndarray,
-    time: np.ndarray,
-    coord_system: Optional[str] = "spherical",
-    difference_scheme: Optional[str] = "forward",
-    time_axis: Optional[int] = -1,
-) -> Tuple[xr.DataArray, xr.DataArray]:
-    """Compute velocity from arrays of positions and time.
-
-    x and y can be provided as longitude and latitude in degrees if
-    coord_system == "spherical" (default), or as easting and northing if
-    coord_system == "cartesian".
-
-    The units of the result are meters per unit of time if
-    coord_system == "spherical". For example, if the time is provided in the
-    units of seconds, the resulting velocity is in the units of meters per
-    second. Otherwise, if coord_system == "cartesian", the units of the
-    resulting velocity correspond to the units of the input. For example,
-    if zonal and meridional displacements are in the units of kilometers and
-    time is in the units of hours, the resulting velocity is in the units of
-    kilometers per hour.
-
-    x, y, and time can be multi-dimensional arrays. If the time axis, along
-    which the finite differencing is performed, is not the last one (i.e.
-    x.shape[-1]), use the time_axis optional argument to specify along which
-    axis should the differencing be done. x, y, and time must have the same
-    shape.
-
-    Difference scheme can take one of three values:
-
-    #. "forward" (default): finite difference is evaluated as ``dx[i] = dx[i+1] - dx[i]``;
-    #. "backward": finite difference is evaluated as ``dx[i] = dx[i] - dx[i-1]``;
-    #. "centered": finite difference is evaluated as ``dx[i] = (dx[i+1] - dx[i-1]) / 2``.
-
-    Forward and backward schemes are effectively the same except that the
-    position at which the velocity is evaluated is shifted one element down in
-    the backward scheme relative to the forward scheme. In the case of a
-    forward or backward difference scheme, the last or first element of the
-    velocity, respectively, is extrapolated from its neighboring point. In the
-    case of a centered difference scheme, the start and end boundary points are
-    evaluated using the forward and backward difference scheme, respectively.
-
-    Parameters
-    ----------
-    x : array_like
-        An N-d array of x-positions (longitude in degrees or zonal displacement in any unit)
-    y : array_like
-        An N-d array of y-positions (latitude in degrees or meridional displacement in any unit)
-    time : array_like
-        An N-d array of times as floating point values (in any unit)
-    coord_system : str, optional
-        Coordinate system that x and y arrays are in; possible values are "spherical" (default) or "cartesian".
-    difference_scheme : str, optional
-        Difference scheme to use; possible values are "forward", "backward", and "centered".
-    time_axis : int, optional
-        Axis along which to differentiate (default is -1)
-
-    Returns
-    -------
-    u : np.ndarray
-        Zonal velocity
-    v : np.ndarray
-        Meridional velocity
-
-    Raises
-    ------
-    ValueError
-        If x and y do not have the same shape.
-        If time_axis is outside of the valid range.
-        If lengths of x, y, and time along time_axis are not equal.
-        If coord_system is not "spherical" or "cartesian".
-        If difference_scheme is not "forward", "backward", or "centered".
-
-    See Also
-    --------
-    :func:`position_from_velocity`
-    """
-
-    # Position arrays must have the same shape.
-    # Although the exception would be raised further down in the function,
-    # we do the check here for a clearer error message.
-    if not x.shape == y.shape:
-        raise ValueError("x and y arrays must have the same shape.")
-
-    # time_axis must be in valid range
-    if time_axis < -1 or time_axis > len(x.shape) - 1:
-        raise ValueError(
-            f"time_axis ({time_axis}) is outside of the valid range ([-1,"
-            f" {len(x.shape) - 1}])."
-        )
-
-    # Input arrays must have the same length along the time axis.
-    if not x.shape[time_axis] == y.shape[time_axis] == time.shape[time_axis]:
-        raise ValueError(
-            f"x, y, and time must have the same length along the time axis "
-            f"({time_axis})."
-        )
-
-    # Swap axes so that we can differentiate along the last axis.
-    # This is a syntax convenience rather than memory access optimization:
-    # np.swapaxes returns a view of the array, not a copy, if the input is a
-    # NumPy array. Otherwise, it returns a copy. For readability, introduce new
-    # variable names so that we can more easily differentiate between the
-    # original arrays and those with swapped axes.
-    x_ = np.swapaxes(x, time_axis, -1)
-    y_ = np.swapaxes(y, time_axis, -1)
-    time_ = np.swapaxes(time, time_axis, -1)
-
-    dx = np.empty(x_.shape)
-    dy = np.empty(y_.shape)
-    dt = np.empty(time_.shape)
-
-    # Compute dx, dy, and dt
-    if difference_scheme == "forward":
-        # All values except the ending boundary value are computed using the
-        # 1st order forward differencing. The ending boundary value is
-        # computed using the 1st order backward difference.
-
-        # Time
-        dt[..., :-1] = np.diff(time_)
-        dt[..., -1] = dt[..., -2]
-
-        # Space
-        if coord_system == "cartesian":
-            dx[..., :-1] = np.diff(x_)
-            dx[..., -1] = dx[..., -2]
-            dy[..., :-1] = np.diff(y_)
-            dy[..., -1] = dy[..., -2]
-
-        elif coord_system == "spherical":
-            distances = distance(y_[..., :-1], x_[..., :-1], y_[..., 1:], x_[..., 1:])
-            bearings = bearing(y_[..., :-1], x_[..., :-1], y_[..., 1:], x_[..., 1:])
-            dx[..., :-1] = distances * np.cos(bearings)
-            dx[..., -1] = dx[..., -2]
-            dy[..., :-1] = distances * np.sin(bearings)
-            dy[..., -1] = dy[..., -2]
-
-        else:
-            raise ValueError('coord_system must be "spherical" or "cartesian".')
-
-    elif difference_scheme == "backward":
-        # All values except the starting boundary value are computed using the
-        # 1st order backward differencing. The starting boundary value is
-        # computed using the 1st order forward difference.
-
-        # Time
-        dt[..., 1:] = np.diff(time_)
-        dt[..., 0] = dt[..., 1]
-
-        # Space
-        if coord_system == "cartesian":
-            dx[..., 1:] = np.diff(x_)
-            dx[..., 0] = dx[..., 1]
-            dy[..., 1:] = np.diff(y_)
-            dy[..., 0] = dy[..., 1]
-
-        elif coord_system == "spherical":
-            distances = distance(y_[..., :-1], x_[..., :-1], y_[..., 1:], x_[..., 1:])
-            bearings = bearing(y_[..., :-1], x_[..., :-1], y_[..., 1:], x_[..., 1:])
-            dx[..., 1:] = distances * np.cos(bearings)
-            dx[..., 0] = dx[..., 1]
-            dy[..., 1:] = distances * np.sin(bearings)
-            dy[..., 0] = dy[..., 1]
-
-        else:
-            raise ValueError('coord_system must be "spherical" or "cartesian".')
-
-    elif difference_scheme == "centered":
-        # Inner values are computed using the 2nd order centered differencing.
-        # The start and end boundary values are computed using the 1st order
-        # forward and backward differencing, respectively.
-
-        # Time
-        dt[..., 1:-1] = (time_[..., 2:] - time_[..., :-2]) / 2
-        dt[..., 0] = time_[..., 1] - time_[..., 0]
-        dt[..., -1] = time_[..., -1] - time_[..., -2]
-
-        # Space
-        if coord_system == "cartesian":
-            dx[..., 1:-1] = (x_[..., 2:] - x_[..., :-2]) / 2
-            dx[..., 0] = x_[..., 1] - x_[..., 0]
-            dx[..., -1] = x_[..., -1] - x_[..., -2]
-            dy[..., 1:-1] = (y_[..., 2:] - y_[..., :-2]) / 2
-            dy[..., 0] = y_[..., 1] - y_[..., 0]
-            dy[..., -1] = y_[..., -1] - y_[..., -2]
-
-        elif coord_system == "spherical":
-            # Inner values
-            y1 = (y_[..., :-2] + y_[..., 1:-1]) / 2
-            x1 = (x_[..., :-2] + x_[..., 1:-1]) / 2
-            y2 = (y_[..., 2:] + y_[..., 1:-1]) / 2
-            x2 = (x_[..., 2:] + x_[..., 1:-1]) / 2
-            distances = distance(y1, x1, y2, x2)
-            bearings = bearing(y1, x1, y2, x2)
-            dx[..., 1:-1] = distances * np.cos(bearings)
-            dy[..., 1:-1] = distances * np.sin(bearings)
-
-            # Boundary values
-            distance1 = distance(y_[..., 0], x_[..., 0], y_[..., 1], x_[..., 1])
-            bearing1 = bearing(y_[..., 0], x_[..., 0], y_[..., 1], x_[..., 1])
-            dx[..., 0] = distance1 * np.cos(bearing1)
-            dy[..., 0] = distance1 * np.sin(bearing1)
-            distance2 = distance(y_[..., -2], x_[..., -2], y_[..., -1], x_[..., -1])
-            bearing2 = bearing(y_[..., -2], x_[..., -2], y_[..., -1], x_[..., -1])
-            dx[..., -1] = distance2 * np.cos(bearing2)
-            dy[..., -1] = distance2 * np.sin(bearing2)
-
-        else:
-            raise ValueError('coord_system must be "spherical" or "cartesian".')
-
-    else:
-        raise ValueError(
-            'difference_scheme must be "forward", "backward", or "centered".'
-        )
-
-    # This should avoid an array copy when returning the result
-    dx /= dt
-    dy /= dt
-
-    return np.swapaxes(dx, time_axis, -1), np.swapaxes(dy, time_axis, -1)
-
-
-def mask_var(
-    var: xr.DataArray,
-    criterion: Union[tuple, list, np.ndarray, xr.DataArray, bool, float, int],
-) -> xr.DataArray:
-    """Return the mask of a subset of the data matching a test criterion.
-
-    Parameters
-    ----------
-    var : xr.DataArray
-        DataArray to be subset by the criterion
-    criterion : array-like
-        The criterion can take three forms:
-        - tuple: (min, max) defining a range
-        - list, np.ndarray, or xr.DataArray: An array-like defining multiples values
-        - scalar: value defining a single value
-
-    Examples
-    --------
-    >>> x = xr.DataArray(data=np.arange(0, 5))
-    >>> mask_var(x, (2, 4))
-    <xarray.DataArray (dim_0: 5)>
-    array([False, False,  True,  True,  True])
-    Dimensions without coordinates: dim_0
-
-    >>> mask_var(x, [0, 2, 4])
-    <xarray.DataArray (dim_0: 5)>
-    array([ True, False, True,  False, True])
-    Dimensions without coordinates: dim_0
-
-    >>> mask_var(x, 4)
-    <xarray.DataArray (dim_0: 5)>
-    array([False, False, False,  True, False])
-    Dimensions without coordinates: dim_0
-
-    Returns
-    -------
-    mask : xr.DataArray
-        The mask of the subset of the data matching the criteria
-    """
-    if isinstance(criterion, tuple):  # min/max defining range
-        mask = np.logical_and(var >= criterion[0], var <= criterion[1])
-    elif isinstance(
-        criterion, (list, np.ndarray, xr.DataArray)
-    ):  # select multiple values
-        # Ensure we define the mask as boolean, otherwise it will inherit
-        # the dtype of the variable which may be a string, object, or other.
-        mask = xr.zeros_like(var, dtype=bool)
-        for v in criterion:
-            mask = np.logical_or(mask, var == v)
-    else:  # select one specific value
-        mask = var == criterion
-    return mask
-
-
 def subset(
     ds: xr.Dataset,
     criteria: dict,
@@ -1102,9 +617,9 @@ def subset(
     for key in criteria.keys():
         if key in ds:
             if ds[key].dims == (traj_dim_name,):
-                mask_traj = np.logical_and(mask_traj, mask_var(ds[key], criteria[key]))
+                mask_traj = np.logical_and(mask_traj, _mask_var(ds[key], criteria[key]))
             elif ds[key].dims == (obs_dim_name,):
-                mask_obs = np.logical_and(mask_obs, mask_var(ds[key], criteria[key]))
+                mask_obs = np.logical_and(mask_obs, _mask_var(ds[key], criteria[key]))
         else:
             raise ValueError(f"Unknown variable '{key}'.")
 
@@ -1134,7 +649,7 @@ def subset(
         return ds_sub
 
 
-def unpack_ragged(
+def unpack(
     ragged_array: np.ndarray,
     rowsize: np.ndarray[int],
     rows: Union[int, Iterable[int]] = None,
@@ -1171,10 +686,10 @@ def unpack_ragged(
 
     .. code-block:: python
 
-        lon = unpack_ragged(ds.lon, ds["rowsize"]) # return a list[xr.DataArray] (slower)
-        lon = unpack_ragged(ds.lon.values, ds["rowsize"]) # return a list[np.ndarray] (faster)
-        first_lon = unpack_ragged(ds.lon.values, ds["rowsize"], rows=0) # return only the first row
-        first_two_lons = unpack_ragged(ds.lon.values, ds["rowsize"], rows=[0, 1]) # return first two rows
+        lon = unpack(ds.lon, ds["rowsize"]) # return a list[xr.DataArray] (slower)
+        lon = unpack(ds.lon.values, ds["rowsize"]) # return a list[np.ndarray] (faster)
+        first_lon = unpack(ds.lon.values, ds["rowsize"], rows=0) # return only the first row
+        first_two_lons = unpack(ds.lon.values, ds["rowsize"], rows=[0, 1]) # return first two rows
 
     Looping over trajectories in a ragged Xarray Dataset to compute velocities
     for each:
@@ -1182,9 +697,9 @@ def unpack_ragged(
     .. code-block:: python
 
         for lon, lat, time in list(zip(
-            unpack_ragged(ds.lon.values, ds["rowsize"]),
-            unpack_ragged(ds.lat.values, ds["rowsize"]),
-            unpack_ragged(ds.time.values, ds["rowsize"])
+            unpack(ds.lon.values, ds["rowsize"]),
+            unpack(ds.lat.values, ds["rowsize"]),
+            unpack(ds.time.values, ds["rowsize"])
         )):
             u, v = velocity_from_position(lon, lat, time)
     """
@@ -1198,3 +713,57 @@ def unpack_ragged(
     unpacked = np.split(ragged_array, indices[1:-1], axis=axis)
 
     return [unpacked[i] for i in rows]
+
+
+def _mask_var(
+    var: xr.DataArray,
+    criterion: Union[tuple, list, np.ndarray, xr.DataArray, bool, float, int],
+) -> xr.DataArray:
+    """Return the mask of a subset of the data matching a test criterion.
+
+    Parameters
+    ----------
+    var : xr.DataArray
+        DataArray to be subset by the criterion
+    criterion : array-like
+        The criterion can take three forms:
+        - tuple: (min, max) defining a range
+        - list, np.ndarray, or xr.DataArray: An array-like defining multiples values
+        - scalar: value defining a single value
+
+    Examples
+    --------
+    >>> x = xr.DataArray(data=np.arange(0, 5))
+    >>> _mask_var(x, (2, 4))
+    <xarray.DataArray (dim_0: 5)>
+    array([False, False,  True,  True,  True])
+    Dimensions without coordinates: dim_0
+
+    >>> _mask_var(x, [0, 2, 4])
+    <xarray.DataArray (dim_0: 5)>
+    array([ True, False, True,  False, True])
+    Dimensions without coordinates: dim_0
+
+    >>> _mask_var(x, 4)
+    <xarray.DataArray (dim_0: 5)>
+    array([False, False, False,  True, False])
+    Dimensions without coordinates: dim_0
+
+    Returns
+    -------
+    mask : xr.DataArray
+        The mask of the subset of the data matching the criteria
+    """
+    if isinstance(criterion, tuple):  # min/max defining range
+        mask = np.logical_and(var >= criterion[0], var <= criterion[1])
+    elif isinstance(
+        criterion, (list, np.ndarray, xr.DataArray)
+    ):  # select multiple values
+        # Ensure we define the mask as boolean, otherwise it will inherit
+        # the dtype of the variable which may be a string, object, or other.
+        mask = xr.zeros_like(var, dtype=bool)
+        for v in criterion:
+            mask = np.logical_or(mask, var == v)
+    else:  # select one specific value
+        mask = var == criterion
+    return mask
